Apply missing patches from original subversion: [r345] to [r360]

+++ src/common/src/ASCII_stringManipulate.c
@@ -12,7 +12,7 @@
  * @date Sep 3, 2010
  * @author Rumen Kyusakov
  * @version 0.5
- * @par[Revision] $Id: ASCII_stringManipulate.c 328 2013-10-30 16:00:10Z kjussakov $
+ * @par[Revision] $Id$
  */
 
 #include "stringManipulate.h"
@@ -274,7 +274,7 @@ void printString(const String* inStr)
 	if(inStr->length == 0)
 		return;
 
-	DEBUG_OUTPUT(("%.*s", inStr->length, inStr->str));
+	DEBUG_OUTPUT(("%.*s", (int) inStr->length, inStr->str));
 }
 
 #endif /* EXIP_DEBUG */
